test/search: add check for slightly tricky search
It's not really tricky, there are many addresses with a '-' in them. I
personally really want to differentiate between foo@bar.com,
foo-testing@bar.com, and foo-patches@bar.com.

This fails, but it shouldn't:

 FAIL   Search by from (tricky):
	--- search.17.expected	2011-05-04 21:52:52.913218077 +0000
	+++ search.17.output	2011-05-04 21:52:52.913218077 +0000
	@@ -1 +1,3 @@
	+thread:XXX   2000-01-01 [1/1] Search By From Name; search by from (name) (inbox unread)
	+thread:XXX   2000-01-01 [1/1] search-by-from-trick@foo.com; search by from (tricky trick) (inbox unread)
	 thread:XXX   2000-01-01 [1/1] search-by-from@foo.com; search by from (tricky) (inbox unread)

One possibility to make this work is to translate @ to -at- in the
database, so search-by-from@foo.com becomes search-by-from-trick-at-foo.com,
therefore searching for search-by-from-at- would not be confused.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>